Adds SpanStoreStorageWithIndexAdapter to bridge SpanStore to legacy code
The Index and Storage traits need to be removed, but we need to approach
this is steps. Firstly, we need to get the existing Builder approach to
work with SpanStore. That way, we can keep the query and collector
services.

This makes a bridge called `SpanStoreStorageWithIndexAdapter` and ports
redis to use it. Other ports will follow.