Removes TTL from Query API
TTLs in zipkin are backend-specific and at best partially supported as
they aren't exposed in the UI. This change removes TTL from the query
api, and removes trace-specific TTL controls. It leaves support for
global TTLs in Cassandra and Redis.

`getDataTimeToLive` was implemented by Cassandra and Redis, but not used
by the UI. `setTimeToLive` by trace was implemented by Cassandra and
Redis, but silently fails when unsupported. Cassandra support of this
feature led to brittle tests and corresponding complexity. When we
consider TTL by trace wasn't used by the UI either, but rather an
undocumented "pin" api, the feature's existence becomes more curious.
This isn't helped by no history of this feature in any GitHub issue.