Clarifies a trace as a list of spans
`Span.annotations` is a list because it is ordered. A trace is an
ordered list of spans, yet in scala we designate it as a Seq. This led
to numerous bugs where people tried to order seqs. The fixes to those
resulted in all implementations using `List` not `Seq`. This change
formalizes that relationship, which has a side-effect of making trace
transformations more likely to retain order.