Moves clock skew correction to SpanStore, for portable testing
Clock skew correction is vital for the UI to appear properly,
particularly in the current model. This is because there's no concept of
sequence of spans except by timestamp.

The type of clock skew adjustment zipkin does is shifting child spans to
a later timestamp when they appear to happen before their parent.

Formerly, there were no tests to show span stores actually do this. This
means you can create implementations which appear portable, but are not
useful as they result in a poor UI.

This moves the implementation of clock skew correction to the common
package, and introduces a new test in SpanStoreSpec which asserts what
expectations are, without limiting how implementations correct clock
skew. For example, this test will pass regardless of whether clock skew
is corrected at storage or query time.