Converts scala functions Span.timestamp, duration to vals
Before this change, Span.timestamp and duration were always derived
lazily from annotations. This decouples that logic by converting the
scala methods to vals and populating them with
`ApplyTimestampAndDuration`.

This serves us in at least two ways. First, the implicit association
between annotations and timestamp or duration was tested in various
components. Making these explicit centralizes the responsibility, and
lowers the test burden on other components. Also, we want to formalize
these fields in persisted models in support of duration queries and
local spans (#807). Organizing logic ahead of this work makes the change
to persistence simpler.